src/requirements.txt
numpy
pandas
deap
joblib
jmetalpy
scipy
ansible
scikit-learn
onnx